requirements.txt
python-dotenv
discord.py

gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client